+++ modules/carto/src/api/index.js
@@ -1,3 +1,3 @@
 export {FORMATS, MAP_TYPES, API_VERSIONS} from './maps-api-common';
-export {getDataV1, CONNECTIONS} from './maps-client';
-export {getData, mapInstantiation} from './maps-cloud-native-client';
+export {getDataV2, CONNECTIONS} from './maps-client';
+export {getData, mapInstantiation} from './maps-v3-client';
